requirements.txt
requests>=2.31.0
google-genai>=1.0.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
